// src/types/types.d.ts
export interface components {
  schemas: {
// ...
    CreateWineBody: {
      name: string;
      region: string;
      image: string;
      /** Format: double */
      price: number;
      type: components["schemas"]["WineType"];
    };
// ...
    WineDetailType: {
      /** Format: double */
      id: number;
      name: string;
      region: string;
      image: string;
      /** Format: double */
      price: number;
      type: string;
      /** Format: double */
      avgRating: number;
      /** Format: double */
      reviewCount: number;
      recentReview: {
        user: components["schemas"]["ReviewUser"];
        /** Format: date-time */
        updatedAt: string;
        /** Format: date-time */
        createdAt: string;
        content: string;
        aroma: components["schemas"]["Aroma"][];
        /** Format: double */
        rating: number;
        /** Format: double */
        id: number;
      } | null;
      /** Format: double */
      userId: number;
      reviews: components["schemas"]["ReviewListType"][];
      avgRatings: components["schemas"]["Record_number.number_"];
    };
// ...
    ReviewDetailType: {
      /** Format: double */
      id: number;
      /** Format: double */
      rating: number;
      /** Format: double */
      lightBold: number;
      /** Format: double */
      smoothTannic: number;
      /** Format: double */
      drySweet: number;
      /** Format: double */
      softAcidic: number;
      aroma: components["schemas"]["Aroma"][];
      content: string;
      /** Format: date-time */
      createdAt: string;
      /** Format: date-time */
      updatedAt: string;
      user: components["schemas"]["ReviewUser"];
      isLiked?: components["schemas"]["Boolean"];
      /** Format: double */
      wineId: number;
      teamId: string;
    };
    CreateReviewBody: {
      /** Format: double */
      rating: number;
      /** Format: double */
      lightBold: number;
      /** Format: double */
      smoothTannic: number;
      /** Format: double */
      drySweet: number;
      /** Format: double */
      softAcidic: number;
      aroma: components["schemas"]["Aroma"][];
      content: string;
      /** Format: double */
      wineId: number;
    };
    UpdateReviewBody: {
      /** Format: double */
      rating?: number;
      /** Format: double */
      lightBold?: number;
      /** Format: double */
      smoothTannic?: number;
      /** Format: double */
      drySweet?: number;
      /** Format: double */
      softAcidic?: number;
      aroma?: components["schemas"]["Aroma"][];
      content?: string;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}